% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_x13story.R
\name{parse_x13story}
\alias{parse_x13story}
\title{Prepare an .Rmd file for use in the interactive tools}
\usage{
parse_x13story(file)
}
\arguments{
\item{file}{path}
}
\description{
Prepare an .Rmd file for use in the interactive tools
}
