% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify.R
\name{prettify}
\alias{prettify}
\title{Prettier Plots (And Other Stuff, at Some Point)}
\usage{
prettify(expr, family = "Palatino", grid = TRUE, box = FALSE)
}
\arguments{
\item{expr}{An R expression, usually resulting in a plot}

\item{family}{font family, to be matched with latex font}
}
\description{
Prettier Plots (And Other Stuff, at Some Point)
}
\examples{
\dontrun{
library(seasonal)
m <- seas(AirPassengers)

prettify(pacf(resid(m), main = ""))
prettify(plot(spectrum(diff(resid(m)), method = "ar"), main = ""))
prettify(plot(m, main = "", xlab = ""))
prettify(monthplot(m, main = "", xlab = ""))
prettify(plot(AirPassengers))
prettify(plot(density(resid(m)), main = ""))
prettify(qqnorm(resid(m), main = ""))
}
}

