% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyplot.R
\name{prettyplot}
\alias{prettyplot}
\title{A more Tufte-like plot}
\usage{
prettyplot(x, ylab = NULL, family = "Palatino",
  legend.pos = "bottomright", ...)
}
\arguments{
\item{x}{a time series}

\item{legend.pos}{position of the legend}

\item{y}{label (optional)}

\item{font}{family, to be matched with latex font}
}
\description{
A more Tufte-like plot
}

