% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettysummary.R
\name{prettysummary}
\alias{prettysummary}
\title{A PDF summary function for seasonal}
\usage{
prettysummary(x, caption = NULL)
}
\arguments{
\item{x}{an object of class \code{"seas"}}

\item{caption}{character string containing the caption}
}
\description{
A PDF summary function for seasonal
}

