% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{viewer}
\alias{viewer}
\title{Deprecated Viewer Function}
\usage{
viewer(x)
}
\arguments{
\item{x}{character, path to the rmarkdown file containing an X-13 story.}
}
\description{
Deprecated Viewer Function
}

