% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x13_handout.R
\name{x13_handout}
\alias{x13_handout}
\title{X-13 story format.}
\usage{
x13_handout(...)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating X-13 stories.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "x13_handout", package = "x13story")
}

}

