% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x13page.R
\name{x13page}
\alias{x13page}
\title{Initialize an interacitve view}
\usage{
x13page(m, series = "main", pdf = TRUE)
}
\arguments{
\item{m}{an object of class \code{"seas"}}

\item{series}{character string indicating which series to show (see
\code{?series})}

\item{pdf}{logical, should a plot be drawn in pdf mode}
}
\description{
Initialize an interacitve view
}
\examples{
# view(story = system.file("stories", "outlier.Rmd", package="x13story"))
}
