% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x13view.R
\name{x13view}
\alias{x13view}
\title{Initialize an interacitve view}
\usage{
x13view(m, series = "main")
}
\arguments{
\item{m}{an object of class \code{"seas"}}

\item{caption}{character string indicating which series to show (see
\code{?series})}
}
\description{
Initialize an interacitve view
}

