% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettysummary.R
\name{prettysummary}
\alias{prettysummary}
\title{A Prettier Summary}
\usage{
prettysummary(x, caption = NULL)
}
\arguments{
\item{x}{an object of class \code{seas}.}

\item{caption}{table caption (optional)}

\item{...}{further arguments}
}
\description{
To be used in X-13 story \code{.Rmd} files for PDF production.
}

